+++ server/controllers/api/channel/shortId/index.js
@@ -1,5 +1,5 @@
-const { handleErrorResponse } = require('../utils/errorHandlers.js');
-const db = require('../../models/index');
+const { handleErrorResponse } = require('../../../utils/errorHandlers.js');
+const db = require('../../../../models');
 
 /*
