+++ server/controllers/api/claim/availability/index.js
@@ -1,6 +1,6 @@
 const checkClaimAvailability = require('../utils/checkClaimAvailability.js');
-const { sendGATimingEvent } = require('../../utils/googleAnalytics.js');
-const { handleErrorResponse } = require('../utils/errorHandlers.js');
+const { sendGATimingEvent } = require('../../../../utils/googleAnalytics.js');
+const { handleErrorResponse } = require('../../../utils/errorHandlers.js');
 
 /*
