+++ server/controllers/api/claim/publish/deleteFile.js
