+++ server/controllers/api/claim/publish/index.js
@@ -1,15 +1,13 @@
-const { details: { host } } = require('../../../config/siteConfig.js');
-const { authenticateUser } = require('../../auth/authentication.js');
-const { sendGATimingEvent } = require('../../utils/googleAnalytics.js');
-const { handleErrorResponse } = require('../utils/errorHandlers.js');
-const checkClaimAvailability = require('../utils/checkClaimAvailability');
-const publish = require('../utils/publish.js');
-const {
-  createBasicPublishParams,
-  createThumbnailPublishParams,
-  parsePublishApiRequestBody,
-  parsePublishApiRequestFiles,
-} = require('../utils/publishHelpers.js');
+const { details: { host } } = require('../../../../../config/siteConfig.js');
+const { authenticateUser } = require('../../../../auth/authentication.js');
+const { sendGATimingEvent } = require('../../../../utils/googleAnalytics.js');
+const { handleErrorResponse } = require('../../../utils/errorHandlers.js');
+const publish = require('./publish.js');
+const createBasicPublishParams = require('./createBasicPublishParams.js');
+const createThumbnailPublishParams = require('./createThumbnailPublishParams.js');
+const parsePublishApiRequestBody = require('./parsePublishApiRequestBody.js');
+const parsePublishApiRequestFiles = require('./parsePublishApiRequestFiles.js');
+const checkClaimAvailability = require('../utils/checkClaimAvailability.js');
 
 /*
