+++ server/controllers/assets/serveByClaim/index.js
@@ -1,11 +1,11 @@
-const { sendGAServeEvent } = require('../../utils/googleAnalytics');
+const { sendGAServeEvent } = require('../../../utils/googleAnalytics');
 const {
   determineResponseType,
   logRequestData,
   getClaimIdAndServeAsset,
-} = require('../utils/serve.js');
-const lbryUri = require('../utils/lbryUri.js');
-const handleShowRender = require('../../render/build/handleShowRender.js');
+} = require('../../utils/serve.js');
+const lbryUri = require('../../utils/lbryUri.js');
+const handleShowRender = require('../../../render/build/handleShowRender.js');
 const SERVE = 'SERVE';
 
 /*
