+++ server/controllers/assets/serveByIdentifierAndClaim/index.js
@@ -1,12 +1,12 @@
-const { sendGAServeEvent } = require('../../utils/googleAnalytics');
+const { sendGAServeEvent } = require('../../../utils/googleAnalytics');
 const {
   determineResponseType,
   flipClaimNameAndIdForBackwardsCompatibility,
   logRequestData,
   getClaimIdAndServeAsset,
-} = require('../utils/serve.js');
-const lbryUri = require('../utils/lbryUri.js');
-const handleShowRender = require('../../render/build/handleShowRender.js');
+} = require('../../utils/serve.js');
+const lbryUri = require('../../utils/lbryUri.js');
+const handleShowRender = require('../../../render/build/handleShowRender.js');
 
 const SERVE = 'SERVE';
