+++ server/controllers/utils/serveHelpers.js
@@ -1,6 +1,5 @@
-const db = require('../../models/index');
+const db = require('../../models');
 const logger = require('winston');
-const { returnPaginatedChannelClaims } = require('./channelPagination.js');
 
 const NO_CHANNEL = 'NO_CHANNEL';
 const NO_CLAIM = 'NO_CLAIM';
@@ -47,58 +46,6 @@ module.exports = {
             return resolve(NO_CLAIM);
           }
           resolve(longClaimId);
-        })
-        .catch(error => {
-          reject(error);
-        });
-    });
-  },
-  getChannelData (channelName, channelClaimId, page) {
-    return new Promise((resolve, reject) => {
-      // 1. get the long channel Id (make sure channel exists)
-      db.Certificate.getLongChannelId(channelName, channelClaimId)
-        .then(longChannelClaimId => {
-          if (!longChannelClaimId) {
-            return [null, null, null];
-          }
-          // 2. get the short ID and all claims for that channel
-          return Promise.all([longChannelClaimId, db.Certificate.getShortChannelIdFromLongChannelId(longChannelClaimId, channelName)]);
-        })
-        .then(([longChannelClaimId, shortChannelClaimId]) => {
-          if (!longChannelClaimId) {
-            return resolve(NO_CHANNEL);
-          }
-          // 3. return all the channel information
-          resolve({
-            channelName,
-            longChannelClaimId,
-            shortChannelClaimId,
-          });
-        })
-        .catch(error => {
-          reject(error);
-        });
-    });
-  },
-  getChannelClaims (channelName, channelClaimId, page) {
-    return new Promise((resolve, reject) => {
-      // 1. get the long channel Id (make sure channel exists)
-      db.Certificate.getLongChannelId(channelName, channelClaimId)
-        .then(longChannelClaimId => {
-          if (!longChannelClaimId) {
-            return [null, null, null];
-          }
-          // 2. get the short ID and all claims for that channel
-          return Promise.all([longChannelClaimId, db.Claim.getAllChannelClaims(longChannelClaimId)]);
-        })
-        .then(([longChannelClaimId, channelClaimsArray]) => {
-          if (!longChannelClaimId) {
-            return resolve(NO_CHANNEL);
-          }
-          // 3. format the data for the view, including pagination
-          let paginatedChannelViewData = returnPaginatedChannelClaims(channelName, longChannelClaimId, channelClaimsArray, page);
-          // 4. return all the channel information and contents
-          resolve(paginatedChannelViewData);
         })
         .catch(error => {
           reject(error);
