+++ server/routes/api/index.js
@@ -1,16 +1,16 @@
-const channelAvailability = require('../../controllers/api/channelAvailability');
-const channelClaims = require('../../controllers/api/channelClaims');
-const channelData = require('../../controllers/api/channelData');
-const channelShortId = require('../../controllers/api/channelShortId');
-const claimAvailability = require('../../controllers/api/claimAvailability');
-const claimData = require('../../controllers/api/claimData');
-const claimGet = require('../../controllers/api/claimGet');
-const claimLongId = require('../../controllers/api/claimLongId');
-const claimPublish = require('../../controllers/api/claimPublish');
-const claimResolve = require('../../controllers/api/claimResolve');
-const claimShortId = require('../../controllers/api/claimShortId');
-const claimList = require('../../controllers/api/claimList');
-const fileAvailability = require('../../controllers/api/fileAvailability');
+const channelAvailability = require('../../controllers/api/channel/availability');
+const channelClaims = require('../../controllers/api/channel/claims');
+const channelData = require('../../controllers/api/channel/data');
+const channelShortId = require('../../controllers/api/channel/shortId');
+const claimAvailability = require('../../controllers/api/claim/availability');
+const claimData = require('../../controllers/api/claim/data/');
+const claimGet = require('../../controllers/api/claim/get');
+const claimLongId = require('../../controllers/api/claim/longId');
+const claimPublish = require('../../controllers/api/claim/publish');
+const claimResolve = require('../../controllers/api/claim/resolve');
+const claimShortId = require('../../controllers/api/claim/shortId');
+const claimList = require('../../controllers/api/claim/list');
+const fileAvailability = require('../../controllers/api/file/availability');
 
 const multipartMiddleware = require('../utils/multipartMiddleware');
