+++ server/routes/assets/index.js
@@ -1,5 +1,5 @@
-const serveAssetByClaim = require('../../controllers/assets/serveAssetByClaim');
-const serveAssetByIdentifierAndClaim = require('../../controllers/assets/serveAssetByIdentifierAndClaim');
+const serveAssetByClaim = require('../../controllers/assets/serveByClaim');
+const serveAssetByIdentifierAndClaim = require('../../controllers/assets/serveByIdentifierAndClaim');
 
 module.exports = (app, db) => {
   app.get('/:identifier/:claim', serveAssetByIdentifierAndClaim);
